Add Creator Module
- The Creator Module makes code more organized. Its purpose is to store all functions that create DOM objects with jQuery.
- Code improvements: added more documentation to planner.js and specified types for parameter documentation.
