+++ public/src/planner.html
@@ -27,7 +27,7 @@
   <link rel="stylesheet" href="css/planner.css">
 
   <!-- Local Javascript -->
-  <script src="js/planner.js" defer></script>
+  <script src="js/planner.js" type="module" defer></script>
 
 </head>
