+++ public/src/prioritizer.html
@@ -25,7 +25,7 @@
   <link rel="stylesheet" href="css/prioritizer.css">
 
   <!-- Local Javascript -->
-  <script src="js/prioritizer.js" defer></script>
+  <script src="js/prioritizer.js" type="module" defer></script>
 
 </head>
